refactor : WinningLotto 객체 생성
