+++ src/main/java/lotto/controller/LottoAutoController.java
@@ -1,9 +1,6 @@
 package lotto.controller;
 
-import lotto.model.LottoPrice;
-import lotto.model.WinningPrice;
-import lotto.model.WinningResult;
-import lotto.model.LottoModel;
+import lotto.model.*;
 import lotto.view.LottoAutoInputView;
 import lotto.view.LottoAutoResultView;
 
@@ -22,14 +19,10 @@ public class LottoAutoController {
         List<LottoModel> lottos = setLotto(quantity);
         resultView.printLotto(quantity, lottos);
 
-        List<Integer> winningNumbers = convertWinningNumbersToInt(inputView.inputWinningNumbers());
-        int bonusNumber = convertBonusNumberToInt(inputView.inputBonusNumber());
-
-        checkWinningNumbers(winningNumbers);
-        checkBonusNumber(bonusNumber);
+        WinningLotto winningLotto = new WinningLotto(inputView.inputWinningNumbers(), inputView.inputBonusNumber());
 
         WinningResult winningResult = new WinningResult();
-        winningResult.getWinningResult(lottos, winningNumbers, bonusNumber);
+        winningResult.getWinningResult(lottos, winningLotto.getWinningNumbers(), winningLotto.getBonusNumber());
         resultView.outputWinningStatistic(getEarningRate(getEarningPrice(), LottoPrice.getPrice()));
     }
 
@@ -39,37 +32,6 @@ public class LottoAutoController {
             lottos.add(new LottoModel());
         }
         return lottos;
-    }
-
-    public List<Integer> convertWinningNumbersToInt(String winningStringNumbers) {
-        List<Integer> winningNumbers = new ArrayList<>();
-        String[] splitWinningNumbers = winningStringNumbers.split(", ");
-        for (int i = 0; i < splitWinningNumbers.length; i++) {
-            int number = Integer.parseInt(splitWinningNumbers[i]);
-            winningNumbers.add(number);
-        }
-        return winningNumbers;
-    }
-
-    private int convertBonusNumberToInt(String bonusStringNumber) {
-        int bonusNumber = Integer.parseInt(bonusStringNumber);
-        return bonusNumber;
-    }
-
-    private void checkWinningNumbers(List<Integer> winningNumbers) {
-        for (int number : winningNumbers) {
-            checkNumber(number);
-        }
-    }
-
-    private void checkBonusNumber(int bonusNumber) {
-        checkNumber(bonusNumber);
-    }
-
-    private void checkNumber(int number) {
-        if (number < 1 || 45 < number) {
-            throw new IllegalArgumentException("당첨번호는 1부터 45까지의 숫자입니다.");
-        }
     }
 
     private float getEarningRate(int earningPrice, int purchasePrice) {
