+++ src/main/java/lotto/model/WinningLotto.java
@@ -1,0 +1,56 @@
+package lotto.model;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class WinningLotto {
+    List<Integer> winningNumbers;
+    int bonusNumber;
+
+    public WinningLotto(String winningNumbers, String bonusNumber) {
+        this.winningNumbers = convertWinningNumbersToInt(winningNumbers);
+        this.bonusNumber = convertBonusNumberToInt(bonusNumber);
+
+        checkWinningNumbers(this.winningNumbers);
+        checkBonusNumber(this.bonusNumber);
+    }
+
+    public List<Integer> convertWinningNumbersToInt(String winningStringNumbers) {
+        List<Integer> winningNumbers = new ArrayList<>();
+        String[] splitWinningNumbers = winningStringNumbers.split(", ");
+        for (int i = 0; i < splitWinningNumbers.length; i++) {
+            int number = Integer.parseInt(splitWinningNumbers[i]);
+            winningNumbers.add(number);
+        }
+        return winningNumbers;
+    }
+
+    private int convertBonusNumberToInt(String bonusStringNumber) {
+        int bonusNumber = Integer.parseInt(bonusStringNumber);
+        return bonusNumber;
+    }
+
+    private void checkWinningNumbers(List<Integer> winningNumbers) {
+        for (int number : winningNumbers) {
+            checkNumber(number);
+        }
+    }
+
+    private void checkBonusNumber(int bonusNumber) {
+        checkNumber(bonusNumber);
+    }
+
+    private void checkNumber(int number) {
+        if (number < 1 || 45 < number) {
+            throw new IllegalArgumentException("당첨번호는 1부터 45까지의 숫자입니다.");
+        }
+    }
+
+    public List<Integer> getWinningNumbers() {
+        return winningNumbers;
+    }
+
+    public int getBonusNumber() {
+        return bonusNumber;
+    }
+}
